refactor : Device -> View로 네이밍 수정하고 view package로 이동
